refactor: rename TapManagerDataSource to tap-manager-data-source
