+++ packages/app/src/tap/commands/run.ts
@@ -1,5 +1,5 @@
 import { posixify } from '../../paths'
-import { tapManagerDataSource } from '../TapManagerDataSource'
+import { tapManagerDataSource } from '../tap-manager-data-source'
 import { defineCommand, TapCommandError } from './definition'
 import { toSpecListEntry } from './specs-list'
 import type { SpecListEntry } from '../types'
