+++ packages/app/src/tap/commands/specs.ts
@@ -1,4 +1,4 @@
-import { tapManagerDataSource } from '../TapManagerDataSource'
+import { tapManagerDataSource } from '../tap-manager-data-source'
 import { defineCommand } from './definition'
 import { toSpecListEntry } from './specs-list'
 import type { SpecListEntry } from '../types'
