+++ packages/app/src/tap/commands/tests.cy.ts
@@ -1,4 +1,4 @@
-import { tapManagerDataSource } from '../TapManagerDataSource'
+import { tapManagerDataSource } from '../tap-manager-data-source'
 import { TapManager } from '../tap-manager'
 
 const CYPRESS_VERSION = '15.0.0'
