+++ packages/app/src/tap/tap-manager-data-source.ts
@@ -2,8 +2,6 @@ import type { FoundSpec } from '@packages/types'
 
 import type { TapTestsRunner } from './types'
 
-// Every runner-window global a tap command reads (or writes) goes through this
-// one seam, so component tests stub a method here instead of the real globals.
 export const tapManagerDataSource = {
   getRunner (): TapTestsRunner | undefined {
     try {
